backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
flask-limiter==3.5.0